+++ docs/changelog.md
@@ -200,7 +200,7 @@ re-export hidden workspace dependencies from `_dep`.
 - update `CONST!` macro with new arms: `hidden macro_export`, `inline macro_export`.
 - add `location` arms to the `_doc!` macro.
 - update `impl_traits!` macro:
-  - add new arms for: `PartialEq`.
+  - add new arms for: `FromStr`, `PartialEq`.
   - change syntax from a single <gen> group to a double [decl][args] group, to support const generics.
 - update `const_assert!` macro:
   - add new arms: `ne_buf`, `ne_str`.